requirements.txt
asgiref==3.3.1
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
Pillow==8.1.0
pkg-resources==0.0.0
pytz==2020.5
sqlparse==0.4.1
